# ENH: More efficient algorithm for unweighted random choice without replacement