--- a/numpy/random/mtrand/mtrand.pyx
+++ b/numpy/random/mtrand/mtrand.pyx
@@ -1057,7 +1057,11 @@ cdef class RandomState:
 
         """
 
+        cdef npy_intp i, j, pop_size, output_size
+        cdef ndarray[npy_intp, ndim=1] tmp_idx_array
+
         # Format and Verify input
+
         a = np.array(a, copy=False)
         if a.ndim == 0:
             try:
@@ -1093,6 +1097,7 @@ cdef class RandomState:
             size = np.prod(shape, dtype=np.intp)
         else:
             size = 1
+        output_size=size
 
         # Actual sampling
         if replace:
@@ -1130,7 +1135,14 @@ cdef class RandomState:
                     n_uniq += new.size
                 idx = found
             else:
-                idx = self.permutation(pop_size)[:size]
+                # perform a partial Knuth shuffle.
+                tmp_idx_array = np.arange(pop_size)
+                i = 0
+                while i < output_size:
+                    j = rk_interval(pop_size - 1 - i, self.internal_state)+i #rand int from [i, pop_size-1]
+                    tmp_idx_array[i], tmp_idx_array[j] = tmp_idx_array[j], tmp_idx_array[i]
+                    i = i + 1
+                idx = tmp_idx_array[:output_size]
                 if shape is not None:
                     idx.shape = shape