@anirudh2290 2020-05-20
According to @bashtage 's last comment this issue is fixed in randomgen merge. Just tried to do some basic runs to verify performance.

On my machine 

Input Size | Output Size | Before (1.16.2) | After (master)
-- | -- | -- | --
1000 | 1000 | 192 us | 69.9 us
10000 | 1000 | 478 us | 268 us
10000 | 3000 | 482 us | 270 us
10000 | 10000 | 492 us | 290 us

So there is close to 2x to 3x speedup depending on input and output size. 


Can we close this for now or does this PR still have potential for speed improvements? Thoughts @bashtage @staticd-growthecommons