@seberg 2020-05-20
Lets close it. If the new API can be further improved, please open a new PR. I think there was a possible speedup which returns the chosen elements themselves not shuffled (i.e. its a random subset, but the random subset may be ordered).